+++ src/command_modules/azure-cli-role/azure/cli/command_modules/role/custom.py
@@ -507,7 +507,7 @@ def list_application_owners(cmd, identifier):
 
 def add_application_owner(cmd, owner_object_id, identifier):
     graph_client = _graph_client_factory(cmd.cli_ctx)
-    owner_url = _get_owner_url(cmd, owner_object_id)
+    owner_url = _get_owner_url(cmd.cli_ctx, owner_object_id)
     return graph_client.applications.add_owner(_resolve_application(graph_client.applications, identifier), owner_url)
 
 
@@ -526,6 +526,13 @@ def list_sps(client, spn=None, display_name=None, query_filter=None):
         sub_filters.append("startswith(displayName,'{}')".format(display_name))
 
     return client.list(filter=(' and '.join(sub_filters)))
+
+
+def list_owned_objects(client, object_type=None):
+    result = client.list_owned_objects()
+    if object_type:
+        result = [r for r in result if r.object_type.lower() == object_type.lower()]
+    return result
 
 
 def list_users(client, upn=None, display_name=None, query_filter=None):
@@ -570,8 +577,14 @@ def get_user_member_groups(cmd, upn_or_object_id, security_enabled_only=False):
     return [{'objectId': x, 'displayName': stubs.get(x)} for x in results]
 
 
-def create_group(client, display_name, mail_nickname):
-    return client.create(GroupCreateParameters(display_name=display_name, mail_nickname=mail_nickname))
+def create_group(cmd, display_name, mail_nickname):
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    group = graph_client.groups.create(GroupCreateParameters(display_name=display_name,
+                                                             mail_nickname=mail_nickname))
+    # TODO: uncomment once design reviewed with AAD team
+    # _set_owner(cmd.cli_ctx, graph_client, group.object_id, graph_client.groups.add_owner)
+
+    return group
 
 
 def check_group_membership(cmd, client, group_id, member_object_id):  # pylint: disable=unused-argument
@@ -598,7 +611,7 @@ def list_group_owners(cmd, group_id):
 
 def add_group_owner(cmd, owner_object_id, group_id):
     graph_client = _graph_client_factory(cmd.cli_ctx)
-    owner_url = _get_owner_url(cmd, owner_object_id)
+    owner_url = _get_owner_url(cmd.cli_ctx, owner_object_id)
     return graph_client.groups.add_owner(_resolve_group(graph_client.groups, group_id), owner_url)
 
 
@@ -618,10 +631,11 @@ def _resolve_group(client, identifier):
     return identifier
 
 
-def create_application(client, display_name, homepage=None, identifier_uris=None,
+def create_application(cmd, display_name, homepage=None, identifier_uris=None,
                        available_to_other_tenants=False, password=None, reply_urls=None,
                        key_value=None, key_type=None, key_usage=None, start_date=None, end_date=None,
                        oauth2_allow_implicit_flow=None, required_resource_accesses=None, native_app=None):
+    graph_client = _graph_client_factory(cmd.cli_ctx)
     key_creds, password_creds, required_accesses = None, None, None
     if native_app:
         if identifier_uris:
@@ -647,7 +661,7 @@ def create_application(client, display_name, homepage=None, identifier_uris=None
                                                   required_resource_access=required_accesses)
 
     try:
-        result = client.create(app_patch_param)
+        result = graph_client.applications.create(app_patch_param)
     except GraphErrorException as ex:
         if 'insufficient privileges' in str(ex).lower():
             link = 'https://docs.microsoft.com/en-us/azure/azure-resource-manager/resource-group-create-service-principal-portal'  # pylint: disable=line-too-long
@@ -663,49 +677,93 @@ def create_application(client, display_name, homepage=None, identifier_uris=None
             ApplicationUpdateParameters._attribute_map['public_client'] = {'key': 'publicClient', 'type': 'bool'}
         app_patch_param = ApplicationUpdateParameters(identifier_uris=[])
         setattr(app_patch_param, 'public_client', True)
-        client.patch(result.object_id, app_patch_param)
-        result = client.get(result.object_id)
+        graph_client.applications.patch(result.object_id, app_patch_param)
+        result = graph_client.applications.get(result.object_id)
 
     return result
 
 
-def list_granted_application(cmd, identifier):
+def list_permissions(cmd, identifier):
+    # the important and hard part is to tell users which permissions have been granted.
+    # we will due diligence to dig out what matters
+
     graph_client = _graph_client_factory(cmd.cli_ctx)
 
-    # Get the Service Principal ObjectId for the client app
+    # first get the permission grant history
     client_sp_object_id = _resolve_service_principal(graph_client.service_principals, identifier)
-
-    # Get the OAuth2 permissions client app
-    permissions = graph_client.oauth2.get(
+    grant_info = graph_client.oauth2.get(
         filter="clientId eq '{}'".format(client_sp_object_id))  # pylint: disable=no-member
+    grant_histories = grant_info.additional_properties['value']
 
-    return permissions.additional_properties['value']
+    # get original permissions required by the application, we will cross check the history
+    # and mark out granted ones
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    application = show_application(graph_client.applications, identifier)
+    permissions = application.required_resource_access
+    for p in permissions:
+        result = list(graph_client.service_principals.list(
+            filter="servicePrincipalNames/any(c:c eq '{}')".format(p.resource_app_id)))
+        granted_times = 'N/A'
+        if result:
+            granted_times = ', '.join([x['startTime'] for x in grant_histories if
+                                       x['resourceId'] == result[0].object_id])
+        setattr(p, 'grantedTime', granted_times)
+    return permissions
 
 
-def grant_application(cmd, identifier, app_id, expires='1'):
+def add_permission(cmd, identifier, api, api_permissions):
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    application = show_application(graph_client.applications, identifier)
+    existing = application.required_resource_access
+    resource_accesses = []
+    for e in api_permissions:
+        access_id, access_type = e.split('=')
+        resource_accesses.append(ResourceAccess(id=access_id, type=access_type))
+    required_resource_access = RequiredResourceAccess(resource_app_id=api,
+                                                      resource_access=resource_accesses)
+    existing.append(required_resource_access)
+    update_parameter = ApplicationUpdateParameters(required_resource_access=existing)
+    graph_client.applications.patch(application.object_id, update_parameter)
+    logger.warning('Invoking "az ad app permission grant --id %s --api %s" is needed to make the '
+                   'change effective', identifier, api)
+
+
+def delete_permission(cmd, identifier, api):
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    application = show_application(graph_client.applications, identifier)
+    existing_accesses = application.required_resource_access
+    existing_accesses = [e for e in existing_accesses if e.resource_app_id != api]
+    update_parameter = ApplicationUpdateParameters(required_resource_access=existing_accesses)
+    return graph_client.applications.patch(application.object_id, update_parameter)
+
+
+def grant_application(cmd, identifier, api, expires='1', scope='user_impersonation'):
     graph_client = _graph_client_factory(cmd.cli_ctx)
 
     # Get the Service Principal ObjectId for the client app
     client_sp_object_id = _resolve_service_principal(graph_client.service_principals, identifier)
 
     # Get the Service Principal ObjectId for associated app
-    associated_sp_object_id = _resolve_service_principal(graph_client.service_principals, app_id)
+    associated_sp_object_id = _resolve_service_principal(graph_client.service_principals, api)
 
     # Build payload
     start_date = datetime.datetime.utcnow()
     end_date = start_date + relativedelta(years=1)
 
-    if expires == '2':
-        end_date = start_date + relativedelta(years=2)
-    elif expires.lower() == 'never':
+    if expires.lower() == 'never':
         end_date = start_date + relativedelta(years=1000)
+    else:
+        try:
+            end_date = start_date + relativedelta(years=int(expires))
+        except ValueError:
+            raise CLIError('usage error: --expires <INT>|never')
 
     payload = {
         "odata.type": "Microsoft.DirectoryServices.OAuth2PermissionGrant",
         "clientId": client_sp_object_id,
         "consentType": "AllPrincipals",
         "resourceId": associated_sp_object_id,
-        "scope": "user_impersonation",
+        "scope": scope,
         "startTime": start_date.isoformat(),
         "expiryTime": end_date.isoformat()
     }
@@ -884,59 +942,42 @@ def list_service_principal_owners(cmd, identifier):
 
 
 def list_service_principal_credentials(cmd, identifier, cert=False):
-    client = _graph_client_factory(cmd.cli_ctx)
-    sp_object_id = _resolve_service_principal(client.service_principals, identifier)
-    app_object_id = _get_app_object_id_from_sp_object_id(client, sp_object_id)
-    sp_creds, app_creds = [], []
-    if cert:
-        sp_creds = list(client.service_principals.list_key_credentials(sp_object_id))
-        if app_object_id:
-            app_creds = list(client.applications.list_key_credentials(app_object_id))
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    if " sp " in cmd.name:
+        sp_object_id = _resolve_service_principal(graph_client.service_principals, identifier)
+        app_object_id = _get_app_object_id_from_sp_object_id(graph_client, sp_object_id)
     else:
-        sp_creds = list(client.service_principals.list_password_credentials(sp_object_id))
-        if app_object_id:
-            app_creds = list(client.applications.list_password_credentials(app_object_id))
+        app_object_id = _resolve_application(graph_client.applications, identifier)
+    return _get_service_principal_credentials(graph_client, app_object_id, cert)
 
-    for x in sp_creds:
-        setattr(x, 'source', 'ServicePrincipal')
-    for x in app_creds:
-        setattr(x, 'source', 'Application')
-    return app_creds + sp_creds
+
+def _get_service_principal_credentials(graph_client, app_object_id, cert=False):
+    if cert:
+        app_creds = list(graph_client.applications.list_key_credentials(app_object_id))
+    else:
+        app_creds = list(graph_client.applications.list_password_credentials(app_object_id))
+
+    return app_creds
 
 
 def delete_service_principal_credential(cmd, identifier, key_id, cert=False):
-    client = _graph_client_factory(cmd.cli_ctx)
-    sp_object_id = _resolve_service_principal(client.service_principals, identifier)
-    if cert:
-        result = list(client.service_principals.list_key_credentials(sp_object_id))
+    graph_client = _graph_client_factory(cmd.cli_ctx)
+    if " sp " in cmd.name:
+        sp_object_id = _resolve_service_principal(graph_client.service_principals, identifier)
+        app_object_id = _get_app_object_id_from_sp_object_id(graph_client, sp_object_id)
     else:
-        result = list(client.service_principals.list_password_credentials(sp_object_id))
+        app_object_id = _resolve_application(graph_client.applications, identifier)
+    result = _get_service_principal_credentials(graph_client, app_object_id, cert)
 
     to_delete = next((x for x in result if x.key_id == key_id), None)
-
-    # we will try to delete the creds at service principal level, if not found, we try application level
-
     if to_delete:
         result.remove(to_delete)
         if cert:
-            return client.service_principals.update_key_credentials(sp_object_id, result)
-        return client.service_principals.update_password_credentials(sp_object_id, result)
+            return graph_client.applications.update_key_credentials(app_object_id, result)
+        return graph_client.applications.update_password_credentials(app_object_id, result)
     else:
-        app_object_id = _get_app_object_id_from_sp_object_id(client, sp_object_id)
-        if app_object_id:
-            if cert:
-                result = list(client.applications.list_key_credentials(app_object_id))
-            else:
-                result = list(client.applications.list_password_credentials(app_object_id))
-            to_delete = next((x for x in result if x.key_id == key_id), None)
-            if to_delete:
-                result.remove(to_delete)
-                if cert:
-                    return client.applications.update_key_credentials(app_object_id, result)
-                return client.applications.update_password_credentials(app_object_id, result)
-
-    raise CLIError("'{}' doesn't exist in the service principal of '{}' or associated application".format(
-        key_id, identifier))
+        raise CLIError("'{}' doesn't exist in the service principal of '{}' or associated application".format(
+            key_id, identifier))
 
 
 def _resolve_service_principal(client, identifier):
@@ -1013,9 +1054,7 @@ def _validate_app_dates(app_start_date, app_end_date, cert_start_date, cert_end_
 # pylint: disable=inconsistent-return-statements
 def create_service_principal_for_rbac(
         # pylint:disable=too-many-statements,too-many-locals, too-many-branches
-        cmd, name=None, password=None, years=None,
-        create_cert=False, cert=None,
-        scopes=None, role='Contributor',
+        cmd, name=None, password=None, years=None, create_cert=False, cert=None, scopes=None, role='Contributor',
         show_auth_for_sdk=None, skip_assignment=False, keyvault=None):
     import time
 
@@ -1056,7 +1095,7 @@ def create_service_principal_for_rbac(
     app_start_date, app_end_date, cert_start_date, cert_end_date = \
         _validate_app_dates(app_start_date, app_end_date, cert_start_date, cert_end_date)
 
-    aad_application = create_application(graph_client.applications,
+    aad_application = create_application(cmd,
                                          display_name=app_display_name,
                                          homepage='https://' + app_display_name,
                                          identifier_uris=[name],
@@ -1067,6 +1106,7 @@ def create_service_principal_for_rbac(
                                          end_date=app_end_date)
     # pylint: disable=no-member
     app_id = aad_application.app_id
+
     # retry till server replication is done
     for l in range(0, _RETRY_TIMES):
         try:
@@ -1084,6 +1124,9 @@ def create_service_principal_for_rbac(
                                                          'response') else ex)  # pylint: disable=no-member
                 raise
     sp_oid = aad_sp.object_id
+
+    # TODO: uncomment once design reviewed with AAD team
+    # _set_owner(cmd.cli_ctx, graph_client, aad_application.object_id, graph_client.applications.add_owner)
 
     # retry while server replication is done
     if not skip_assignment:
@@ -1128,6 +1171,13 @@ def create_service_principal_for_rbac(
             cert_file)
         result['fileWithCertAndPrivateKey'] = cert_file
     return result
+
+
+def _get_signed_in_user_object_id(graph_client):
+    try:
+        return graph_client.signed_in_user.get().object_id
+    except GraphErrorException:  # error could be possible if you logged in as a service principal
+        pass
 
 
 def _get_keyvault_client(cli_ctx):
@@ -1399,14 +1449,20 @@ def _get_object_stubs(graph_client, assignees):
     return result
 
 
-def _get_owner_url(cmd, owner_object_id):
+def _get_owner_url(cli_ctx, owner_object_id):
     if '://' in owner_object_id:
         return owner_object_id
-    graph_url = cmd.cli_ctx.cloud.endpoints.active_directory_graph_resource_id
+    graph_url = cli_ctx.cloud.endpoints.active_directory_graph_resource_id
     from azure.cli.core._profile import Profile
-    profile = Profile(cli_ctx=cmd.cli_ctx)
+    profile = Profile(cli_ctx=cli_ctx)
     _, _2, tenant_id = profile.get_login_credentials()
     return graph_url + tenant_id + '/directoryObjects/' + owner_object_id
+
+
+def _set_owner(cli_ctx, graph_client, asset_object_id, setter):
+    signed_in_user_object_id = _get_signed_in_user_object_id(graph_client)
+    if signed_in_user_object_id:
+        setter(asset_object_id, _get_owner_url(cli_ctx, signed_in_user_object_id))
 
 
 # for injecting test seams to produce predicatable role assignment id for playback