+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -7,6 +7,9 @@ Release History
 ++++++
 * `image create`: expose storage-sku argument for setting the image's default storage account type
 * `vm resize`: fix bug where `--no-wait` option causes command to crash
+* `vm encryption show`: table output format shows status
+* `vm secret format`: requires json/jsonc output. Warns user and defaults to json output if an undesired output format is selected.
+* `vm create --image`: improved validation of image argument
 
 2.2.6
 ++++++
@@ -16,7 +19,7 @@ Release History
 
 2.2.5
 ++++++
-* Fix SDK issue that caused Homebrew instllation to fail.
+* Fix SDK issue that caused Homebrew installation to fail.
 
 2.2.4
 ++++++